requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Flask==3.1.0
flask-mysql-connector==1.1.0
Flask-OAuthlib==0.9.6
Flask-Session==0.8.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msal==1.31.1
msgspec==0.19.0
mysql-connector-python==9.2.0
numpy==2.2.3
oauthlib==2.1.0
pandas==2.2.3
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==1.1.0
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
